Removed 'accountId' parameter from b2_list_file_names.
Issue #5: In addition to removing this parameter from the affected
B2 API call, this PR adds support for the additional parameters accepted,
which prove to be super useful for pruning the result set one is able
to receive. Also added a convenient return of the file list. POD was
updated.

Resolves Issue 5.